requirements.txt
Babel==2.9.1
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
Deprecated==1.2.13
Flask==2.0.2
Flask-BabelEx==0.9.4
Flask-Dance==5.1.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
GitHub-Flask==3.2.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.1.1
passlib==1.7.4
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.5.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
speaklater==1.3
SQLAlchemy==1.4.29
urllib3==1.26.8
URLObject==2.4.3
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==3.0.1
